fix: use int instead of size_t for matrix dimensions

+++ src/matrix.cpp
@@ -3,13 +3,13 @@
 
 // Initialize a new 2d matrix with these values.
 TMath::Matrix::Matrix(std::initializer_list<std::initializer_list<DOUBLE>> list) {
-	size_t height = list.size(), index = 0;
+	int height = list.size(), index = 0;
 	if (height < 1) throw TMath::EMPTY_MATRIX_ERROR;
 
 	elements = std::vector<std::vector<DOUBLE>>(height);
 	for (auto& row : list) elements[index++] = std::vector<DOUBLE>(row);
 
-	size_t common = elements[index-1].size();
+	int common = elements[index-1].size();
 	for (auto& row : elements) if (row.size() != common) throw TMath::DIMENSION_ERROR;
 }
 
@@ -18,23 +18,23 @@ TMath::Matrix::Matrix(const int& width, const int& height) {
 	if (width < 1 || height < 1) throw TMath::EMPTY_MATRIX_ERROR;
 
 	elements = std::vector<std::vector<DOUBLE>>(height);
-	for (size_t index = 0; index < width; index++) {
+	for (int index = 0; index < width; index++) {
 		elements[index] = std::vector<DOUBLE>(width);
 	}
 }
 
 // Initialize a new matrix as a copy of the matrix m.
 TMath::Matrix::Matrix(const Matrix& m) {
-	size_t height = m.elements.size();
+	int height = m.elements.size();
 	elements = std::vector<std::vector<DOUBLE>>(m.elements.size());
-	for (size_t index = 0; index < height; index++) elements[index] = std::vector<DOUBLE>(m.elements[index]);
+	for (int index = 0; index < height; index++) elements[index] = std::vector<DOUBLE>(m.elements[index]);
 }
 
 // Check if the two matrices have the same dimensions.
-std::pair<size_t, size_t> TMath::Matrix::validate(const Matrix& m) const {
-	size_t w = width(), h = height();
+std::pair<int, int> TMath::Matrix::validate(const Matrix& m) const {
+	int w = width(), h = height();
 	if (m.width() != w || m.height() != h) throw TMath::DIMENSION_ERROR;
-	return std::pair<size_t, size_t>(w, h);
+	return std::pair<int, int>(w, h);
 }
 
 // Access a matrix row.
@@ -44,9 +44,9 @@ std::vector<TMath::DOUBLE>& TMath::Matrix::operator[](const int& i) {
 
 // Test if the matrices are equal or nearly equal as specificed by the epsilon.
 bool TMath::Matrix::equal(const Matrix& m, const DOUBLE& eps) const {
-	std::pair<size_t, size_t> dimensions = validate(m);
+	std::pair<int, int> dimensions = validate(m);
 
-	for (size_t i = 0; i < dimensions.second; i++) for (size_t j = 0; j < dimensions.first; j++) 
+	for (int i = 0; i < dimensions.second; i++) for (int j = 0; j < dimensions.first; j++) 
 		if (!TMath::equal(elements[i][j], m.elements[i][j], eps)) return false;
 
 	return true;
@@ -63,23 +63,23 @@ bool TMath::Matrix::operator!=(const Matrix& m) const {
 }
 
 // Get the matrix col count.
-size_t TMath::Matrix::width() const {
+int TMath::Matrix::width() const {
 	return elements[0].size();
 }
 
 // Get the matrix row count.
-size_t TMath::Matrix::height() const {
+int TMath::Matrix::height() const {
 	return elements.size();
 }
 
 // Add two matrices.
 TMath::Matrix TMath::Matrix::operator+(const Matrix& a) const {
-	size_t w = width(), h = height();
+	int w = width(), h = height();
 	if (a.width() != w || a.height() != h) throw TMath::DIMENSION_ERROR;
 
 	Matrix result(a);
-	for (size_t i = 0; i < h; i++) {
-		for (size_t j = 0; j < w; j++) {
+	for (int i = 0; i < h; i++) {
+		for (int j = 0; j < w; j++) {
 			result[i][j] += elements[i][j];
 		}
 	}
@@ -89,12 +89,12 @@ TMath::Matrix TMath::Matrix::operator+(const Matrix& a) const {
 
 // Subtract two matrices.
 TMath::Matrix TMath::Matrix::operator-(const Matrix& a) const {
-	size_t w = width(), h = height();
+	int w = width(), h = height();
 	if (a.width() != w || a.height() != h) throw TMath::DIMENSION_ERROR;
 
 	Matrix result(*this);
-	for (size_t i = 0; i < h; i++) {
-		for (size_t j = 0; j < w; j++) {
+	for (int i = 0; i < h; i++) {
+		for (int j = 0; j < w; j++) {
 			result[i][j] -= a.elements[i][j];
 		}
 	}
@@ -104,11 +104,11 @@ TMath::Matrix TMath::Matrix::operator-(const Matrix& a) const {
 
 // Generate a string representation of the matrix.
 std::string TMath::Matrix::to_string() const {
-	size_t w = width(), h = height();
+	int w = width(), h = height();
 	std::stringstream stream;
 	stream << "{[";
-	for (size_t i = 0; i < h; i++) {
-		for (size_t j = 0; j < w-1; j++) {
+	for (int i = 0; i < h; i++) {
+		for (int j = 0; j < w-1; j++) {
 			stream << elements[i][j] << ", ";
 		}
 		stream << elements[i][w-1];
